+++ R/movecost.R
@@ -63,6 +63,21 @@
 #' when calculating accumulated costs.
 #'
 #'
+#'\strong{Alberti's Tobler hiking function modified for animal foraging excursions}:\cr
+#'
+#' \eqn{ (6 * exp(-3.5 * abs(x[adj] + 0.05))) * 0.25 }\cr
+#'
+#' proposed by Gianmarco Alberti;
+#' \strong{see}: Locating potential pastoral foraging routes in Malta through the use of Geographic Information System (in press).\cr
+#' The Tobler’s function has been rescaled to fit animal walking speed during foraging excursions. The distribution of the latter, as empirical data show, turns out to be right-skewed
+#' and to vary along a continuum. It ranges from very low speed values (corresponding to slow grazing activities grazing while walking) to comparatively higher values
+#' (up to about 4.0 km/h) corresponding to travels without grazing directional travel toward feeding stations. In an attempt to find a balance between different published figures,
+#' the function consider 1.5 km/h as the average flock speed, which roughly corresponds to the average speed recorded in some studies.
+#' The figure is considered the typical speed of flocks during excursions in which grazing takes place while walking, which in most situations can be considered a typical form of grazing.
+#' Tobler’s hiking function has been rescaled by a factor of 0.25 to represent the walking pace of a flock instead of humans.
+#' The factor corresponds to the ratio between the flock average speed (1.5 km/h) and the maximum human walking speed (about 6.0 km/h) on a favourable slope.
+#'
+#'
 #'\strong{Relative energetic expenditure cost function}:\cr
 #'
 #' \eqn{ 1 / (tan((atan(abs(x[adj]))*180/pi)*pi/180) / tan (1*pi/180)) }\cr
@@ -116,6 +131,13 @@
 #' the value of 6 used by Van Leusen (as per Herzog 2013).\cr
 #' \strong{Note}: in the returned charts, the cost is transposed from Watts to Megawatts.\cr
 #'
+#' \strong{Llobera-Sluckin's metabolic energy expenditure cost function (in KJ/m)}:\cr
+#'
+#' \eqn{ 1 / (2.635 + (17.37 * abs(x[adj])) + (42.37 * abs(x[adj])^2) - (21.43 * abs(x[adj])^3) + (14.93 * abs(x[adj])^4))  }\cr
+#'
+#' for which \strong{see} Llobera M.-Sluckin T.J. (2007). Zigzagging: Theoretical insights on climbing strategies, Journal of Theoretical Biology 249, 206-217.\cr
+#'
+#'
 #' \strong{Note} that the walking-speed-related cost functions listed above are used as they are, while the other functions are reciprocated.
 #' This is done since "gdistance works with conductivity rather than the more usual approach using costs"; therefore
 #' "we need inverse cost functions" (Nakoinz-Knitter (2016). "Modelling Human Behaviour in Landscapes". New York: Springer, p. 183).
@@ -139,11 +161,13 @@
 #' \strong{ic} uses the Irmischer-Clarke's modified Tobler's hiking function (on-path);
 #' \strong{icofp} uses the Irmischer-Clarke's modified Tobler's hiking function (off-path);
 #' \strong{ug} uses the Uriarte Gonzalez's slope-dependant walking-time cost function;
+#' \strong{alb} uses the Alberti's Tobler hiking function modified for animal foraging excursions;
 #' \strong{ree} uses the relative energetic expenditure cost function;
 #' \strong{hrz} uses the Herzog's metabolic cost function;
 #' \strong{wcs} uses the wheeled-vehicle critical slope cost function;
 #' \strong{p} uses the Pandolf et al.'s metabolic energy expenditure cost function;
-#' \strong{vl} uses the Van Leusen's metabolic energy expenditure cost function (see Details).
+#' \strong{vl} uses the Van Leusen's metabolic energy expenditure cost function.
+#' \strong{ls} uses the Llobera-Sluckin's metabolic energy expenditure cost function (see Details).
 #' @param time time-unit expressed by the accumulated raster and by the isolines if Tobler's and Tobler-related cost functions are used;
 #' 'h' for hour, 'm' for minutes.
 #' @param outp type of output: 'raster' or 'contours' (see Details).
@@ -340,8 +364,30 @@ movecost <- function (dtm, origin, destin=NULL, funct="t", time="h", outp="r", s
     sub.title.lcp.plot <- paste0("LCP(s) and cost distance(s) based on the Pandolf et al.'s metabolic energy expenditure cost function \n cost in Megawatts; parameters: W: ", W, "; L: ", L, "; N: ", N, "; V: ", V, "\nblack dot=start location\n red dot(s)=destination location(s)")
   }
 
+  if(funct=="ls") {
+    #Llobera-Sluckin's metabolic energy expenditure cost function (in KJ/m)
+    cost_function <- function(x){ 1 / (2.635 + (17.37 * abs(x[adj])) + (42.37 * abs(x[adj])^2) - (21.43 * abs(x[adj])^3) + (14.93 * abs(x[adj])^4)) }
+
+    #set the labels to be used within the returned plot
+    main.title <- "Accumulated cost isolines around origin"
+    sub.title <- paste0("Cost based on the Llobera-Sluckin's metabolic energy expenditure cost function")
+    legend.cost <- "energy expenditure cost (KJ/m)"
+    sub.title.lcp.plot <- paste0("LCP(s) and cost distance(s) based on the Llobera-Sluckin's metabolic energy expenditure cost function \n cost in KJ/m \nblack dot=start location\n red dot(s)=destination location(s)")
+  }
+
+  if(funct=="alb") {
+    #ALberti's modified Tobler hiking function, adapted for anomal foraging excursions
+    cost_function <- function(x){(6 * exp(-3.5 * abs(x[adj] + 0.05))) * 0.25}
+
+    #set the labels to be used within the returned plot
+    main.title <- paste0("Walking-time isochrones (in ", time, ") around origin")
+    sub.title <- "Walking-time based on the Alberti's modified Tobler hiking function"
+    legend.cost <- paste0("walking-time (", time,")")
+    sub.title.lcp.plot <- paste0("LCP(s) and walking-time distance(s) based on the Alberti's modified Tobler hiking function (time in ", time, ") \nblack dot=start location\n red dot(s)=destination location(s)")
+  }
+
   #cost calculation for walking-speed-based cost functions
-  if (funct=="t" | funct=="tofp" | funct=="mt" | funct=="ic" | funct=="icofp") {
+  if (funct=="t" | funct=="tofp" | funct=="mt" | funct=="ic" | funct=="icofp" | funct=="alb") {
 
     #restrict the speed calculation to adjacent cells by creating an index for adjacent cells (adj) with the function 'adjacent'
     adj <- raster::adjacent(dtm, cells=1:ncell(dtm), pairs=TRUE, directions=8)
@@ -361,7 +407,7 @@ movecost <- function (dtm, origin, destin=NULL, funct="t", time="h", outp="r", s
   #cost calculation for other types of cost functions;
   #note the Uriarte Gonzalez's slope-dependant walking-time cost function is in this group since (unlike the above functions)
   #it expresses cost as time NOT speed
-  if (funct=="ree" | funct=="hrz" | funct=="wcs" | funct=="vl" | funct=="p" | funct=="ug") {
+  if (funct=="ree" | funct=="hrz" | funct=="wcs" | funct=="vl" | funct=="p" | funct=="ug" | funct=="ls") {
 
     #restrict the cost calculation to adjacent cells by creating an index for adjacent cells (adj) with the function 'adjacent'
     adj <- raster::adjacent(dtm, cells=1:ncell(dtm), pairs=TRUE, directions=8)
@@ -379,8 +425,8 @@ movecost <- function (dtm, origin, destin=NULL, funct="t", time="h", outp="r", s
   accum_final <- gdistance::accCost(Conductance, sp::coordinates(origin))
 
   #if user select the Tobler's, the modified Tobler's, the Irmischer-Clarke's,
-  #or the Uriarte Gonzalez's functions, turn seconds into the user-defined time-scale
-  if (funct=="t" | funct=="tofp" | funct=="mt" | funct=="ic" | funct=="icofp" | funct=="ug") {
+  #the Uriarte Gonzalez's or the Alberti's function, turn seconds into the user-defined time-scale
+  if (funct=="t" | funct=="tofp" | funct=="mt" | funct=="ic" | funct=="icofp" | funct=="ug" | funct=="alb") {
     if (time=="h") {
       #turn seconds into hours
       accum_final <- accum_final / 3600